Parsing protobufs, I do not know how

+++ own_utils/tachiyomi_gzip.h
@@ -1,11 +1,11 @@
 #ifndef CROI_APP_CHIYO_TATSU_TACHIYOMI_GZIP
 #define CROI_APP_CHIYO_TATSU_TACHIYOMI_GZIP
 
-#include <stdio.h>
+#include <stdint.h>
 
 extern unsigned char **input_uncompressed;
 
-size_t tachiyomi_gzip_load(const char *filename,
-                           unsigned char **input_uncompressed);
+uint64_t tachiyomi_gzip_load(const char *filename,
+                             unsigned char **input_uncompressed);
 
 #endif
